+++ test/_env/app/fiori/webapp/Component.js
@@ -1,0 +1,7 @@
+"use strict";
+
+sap.ui.define(["sap/fe/core/AppComponent"], function (AppComponent) {
+  return AppComponent.extend("fiori.Component", {
+    metadata: { manifest: "json" },
+  });
+});